pyproject.toml
[tool.poetry]
name = "pythonassistent"
version = "0.1.0"
description = ""
authors = ["ghert213 <pavlik.kovalenko.99@list.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.4.1"
assemblyai = "^0.22.0"
ticktick-py = "^2.0.3"
rutimeparser = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
